requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
flask>=2.0.1
python-dotenv>=0.19.0
requests>=2.26.0
aiohttp>=3.7.4
websockets>=9.1
python-telegram-bot==21.4
pandas==2.2.3
numpy==1.26.4
ta==0.11.0
ccxt>=4.3.0
scikit-learn==1.5.2
joblib==1.4.2
polars==1.12.0
schedule==1.2.2
psutil==6.1.0
httpx==0.27.2
cachetools==5.5.0
pytz==2024.1